move phaser to game page
