+++ client/src/pages/Game.js
@@ -1,6 +1,8 @@
 import React from "react";
+import GameSetup from "../components/GameSetup";
 
 function Game() {
+    GameSetup();
     return (
         <h1>Game page</h1>
     );
